Converted some of the math floats to doubles to increase precision. This increases calculation time and work memory however it will be worth it i think...

+++ math/Point2D.cpp
@@ -1,12 +1,12 @@
 #include "Point2D.h"
 #include <math.h>
 
-const float Point2D::getAngleToOrigin()
+const double Point2D::getAngleToOrigin()
 {
   return getAlpha();
 }
 
-const float Point2D::getAngleBetweenPoints(const Point2D &i_point2D) const
+const double Point2D::getAngleBetweenPoints(const Point2D &i_point2D) const
 {
   return ((*this) -i_point2D).getAlpha();
 }
